@@ -4,12 +4,9 @@
   "description": "A cache support for memoized functions (cache-manager adapter)",
   "main": "index.js",
   "scripts": {
-    "test": "mocha tests/**/*.js",
+    "test": "mocha",
     "watch": "npm run test -- -w",
-    "lint": "./node_modules/.bin/eslint --ext .js ./src ./tests",
-    "release:major": "./node_modules/.bin/npm-release major",
-    "release:minor": "./node_modules/.bin/npm-release minor",
-    "release:patch": "./node_modules/.bin/npm-release patch",
+    "lint": "eslint . --fix",
     "precommit": "npm run lint",
     "prepush": "npm run test"
   },
@@ -24,16 +21,18 @@
     "cache-manager": "^1.4.0",
     "cache-manager-redis": "^0.2.2",
     "chai": "^1.10.0",
-    "eslint": "^1.10.3",
+    "eslint": "^5.8.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^8.0.0",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-standard": "^4.0.0",
     "husky": "^0.10.2",
     "lzma-purejs": "^0.9.3",
-    "mocha": "^2.1.0",
-    "npm-release": "^1.0.0"
+    "mocha": "^5.2.0"
   },
   "dependencies": {
-    "async-deco": "^7.6.0",
-    "md5-o-matic": "^0.1.1",
-    "memoize-cache": "^5.0.0"
+    "memoize-cache": "^5.0.3"
   },
   "optionalDependencies": {
     "snappy": "^6.0.4"
